requirements.txt
ultralytics>=8.0.20,<9.0.0
opencv-python>=4.8.0,<5.0.0
pytesseract>=0.3.10,<0.4.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
python-dateutil>=2.8.2,<3.0.0
psycopg2-binary>=2.9.5,<3.0.0
python-dotenv>=1.0.0,<2.0.0
